Module google_transcoder1_beta1::api

Structs

AdBreak
Ad break.

Aes128Encryption
Configuration for AES-128 encryption.

Animation
Animation types.

AnimationEnd
End previous overlay animation from the video. Without AnimationEnd, the overlay object will keep the state of previous animation until the end of the video.

AnimationFade
Display overlay object with fade animation.

AnimationStatic
Display static overlay object.

Audio
Audio preprocessing configuration.

AudioAtom
The mapping for the Job.edit_list atoms with audio EditAtom.inputs.

AudioChannel
The audio channel.

AudioChannelInput
Identifies which input file, track, and channel should be used.

AudioStream
Audio stream resource.

Color
Color preprocessing configuration.

Crop
Video cropping configuration for the input video. The cropped input video is scaled to match the output resolution.

Deblock
Deblock preprocessing configuration.

Denoise
Denoise preprocessing configuration.

EditAtom
Edit atom.

ElementaryStream
Encoding of an input file such as an audio, video, or text track. Elementary streams must be packaged before mapping and sharing between different output formats.

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

Encryption
Encryption settings.

FailureDetail
Additional information about the reasons for the failure.

Image
Overlaid jpeg image.

Input
Input asset.

Job
Transcoding job resource.

JobConfig
Job configuration

JobTemplate
Transcoding job template resource.

ListJobTemplatesResponse
Response message for TranscoderService.ListJobTemplates.

ListJobsResponse
Response message for TranscoderService.ListJobs.

Manifest
Manifest configuration.

MpegCommonEncryption
Configuration for MPEG Common Encryption (MPEG-CENC).

MuxStream
Multiplexing settings for output stream.

NormalizedCoordinate
2D normalized coordinates. Default: {0.0, 0.0}

OriginUri
The origin URI.

Output
Location of output file(s) in a Cloud Storage bucket.

Overlay
Overlay configuration.

Pad
Pad filter configuration for the input video. The padded input video is scaled after padding with black to match the output resolution.

PreprocessingConfig
Preprocessing configurations.

Progress
Estimated fractional progress for each step, from 0 to 1.

ProjectLocationJobCreateCall
Creates a job in the specified region.

ProjectLocationJobDeleteCall
Deletes a job.

ProjectLocationJobGetCall
Returns the job data.

ProjectLocationJobListCall
Lists jobs in the specified region.

ProjectLocationJobTemplateCreateCall
Creates a job template in the specified region.

ProjectLocationJobTemplateDeleteCall
Deletes a job template.

ProjectLocationJobTemplateGetCall
Returns the job template data.

ProjectLocationJobTemplateListCall
Lists job templates in the specified region.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the Transcoder hub.

PubsubDestination
A Pub/Sub destination.

SampleAesEncryption
Configuration for SAMPLE-AES encryption.

SegmentSettings
Segment settings for "ts", "fmp4" and "vtt".

SpriteSheet
Sprite sheet configuration.

TextAtom
The mapping for the Job.edit_list atoms with text EditAtom.inputs.

TextInput
Identifies which input file and track should be used.

TextStream
Encoding of a text stream. For example, closed captions or subtitles.

Transcoder
Central instance to access all Transcoder related resource activities

VideoStream
Video stream resource.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.